#1020: refactor tests by feature
